Setup script

+++ scripts/setup_target.sh
@@ -1,17 +1,24 @@
 #!/usr/bin/env bash
+
+NVM_USER=sam
+
+# Run as root
 echo "Running on $HOSTNAME"
-sudo su -
+if [[ $(id -u) -ne 0 ]] ; then
+  echo "Please run as root"
+  exit 1
+fi
 
 # Headless mode
-echo "$HOSTNAME: Disabling graphics..."
+echo "-- $HOSTNAME: Disabling graphics..."
 systemctl stop gdm3
 systemctl disable gdm3
 systemctl set-default multi-user.target
 
 # Create "sam" user
-if id "$1" &>/dev/null; then
-  echo "$HOSTNAME: User already exists, skipping..."
+if id "$NVM_USER" &>/dev/null; then
+  echo "-- $HOSTNAME: User already exists, skipping..."
 else
-  echo "$HOSTNAME: Creating user..."
-  adduser "$USER"
+  echo "-- $HOSTNAME: Creating user..."
+  adduser "$NVM_USER"
 fi
